requirements.txt
# Backend dependencies — GreenTrace Phase 1 + 2

Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0

# Phase 2
celery>=5.3
redis>=5.0
django-celery-beat>=2.6
